+++ src/zeroband/inference/eval/config.py
@@ -7,8 +7,7 @@ from zeroband.inference import envs
 
 from pydantic import Field
 
-from zeroband.inference.config import LogConfig, ModelConfig, SamplingConfig as InferenceSamplingConfig
-from zeroband.inference.eval.registry import Benchmark
+from zeroband.inference.config import ModelConfig, EvalConfig, LogConfig, SamplingConfig as InferenceSamplingConfig
 from zeroband.utils.config import MultiMonitorConfig
 from zeroband.utils.pydantic_config import BaseConfig, BaseSettings
 
@@ -34,47 +33,6 @@ class ParallelConfig(BaseConfig):
     ]
 
 
-class OnlineEvalConfig(BaseConfig):
-    """Configures online evaluation."""
-
-    ckpt_path: Annotated[
-        Path,
-        Field(
-            default=Path("checkpoints"),
-            description="Path to read checkpoints from when doing online evaluation. Expects subdirectories named 'step_x' within the directory.",
-        ),
-    ]
-
-    interval: Annotated[
-        int,
-        Field(
-            default=100,
-            ge=0,
-            description="Interval at which to evaluate the model.",
-        ),
-    ]
-
-    max_steps: Annotated[
-        int | None,
-        Field(
-            default=None,
-            description="Maximum number of steps to run online evaluation for. If None, will run indefinitely.",
-        ),
-    ]
-
-
-class EvalConfig(BaseConfig):
-    """Configures evaluation."""
-
-    benchmarks: Annotated[
-        list[Benchmark],
-        Field(
-            default=["math500"],
-            description="Benchmarks to evaluate on. By default, it will evaluate only on the MATH-500 benchmark.",
-        ),
-    ]
-
-    online: Annotated[OnlineEvalConfig | None, Field(default=None)]
 
 
 class Config(BaseSettings):
